fix: 매칭 생성 로직 변경
