+++ src/main/java/com/bookbla/americano/domain/member/repository/custom/impl/MemberRepositoryCustomImpl.java
@@ -11,24 +11,14 @@ import com.bookbla.americano.domain.member.enums.MemberStatus;
 import com.bookbla.americano.domain.member.enums.SmokeType;
 import com.bookbla.americano.domain.member.repository.custom.MemberRepositoryCustom;
 import com.bookbla.americano.domain.member.repository.entity.*;
-import com.bookbla.americano.domain.member.service.dto.MemberRecommendationDto;
-import com.bookbla.americano.domain.postcard.repository.entity.Postcard;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
 
-import static com.bookbla.americano.domain.member.repository.entity.QMember.member;
-import static com.bookbla.americano.domain.member.repository.entity.QMemberBook.memberBook;
 import static com.bookbla.americano.domain.school.repository.entity.QSchool.school;
 
 @Repository
@@ -36,8 +26,6 @@ import static com.bookbla.americano.domain.school.repository.entity.QSchool.scho
 public class MemberRepositoryCustomImpl implements MemberRepositoryCustom {
 
     private final JPAQueryFactory queryFactory;
-
-    private static final int MAX_RANDOM_MATCHING_COUNT = 10;
 
     @Override
     public List<BookProfileResponse> getAllMembers(Long memberId, MemberBookProfileRequestDto requestDto) {
@@ -88,31 +76,6 @@ public class MemberRepositoryCustomImpl implements MemberRepositoryCustom {
                 )
                 .orderBy(member.createdAt.desc())
                 .fetch();
-    }
-
-    @Override
-    public List<Map<Long, Long>> getRecommendationMemberIdsAndBookIds(MemberRecommendationDto recommendationDto, List<Postcard> postcards) {
-        Set<Long> excludeMemberIds = recommendationDto.getExcludeMemberIds();
-
-        LocalDateTime twoWeeksAgo = LocalDateTime.now().minusDays(14);
-
-        return queryFactory
-                .selectDistinct(member.id, memberBook.id)
-                .from(memberBook)
-                .join(member).on(memberBook.member.eq(member))
-                .where(
-                        member.id.ne(recommendationDto.getMemberId()),
-                        member.memberProfile.gender.ne(Gender.valueOf(recommendationDto.getMemberGender())),
-                        member.lastUsedAt.coalesce(LocalDate.parse("1900-01-01").atStartOfDay()).after(twoWeeksAgo),
-                        member.id.notIn(excludeMemberIds)
-                ).fetch()
-                .stream()
-                .map(m -> {
-                    Map<Long, Long> map = new HashMap<>();
-                    map.put(m.get(member.id), m.get(memberBook.id));
-                    return map;
-                })
-                .collect(Collectors.toList());
     }
 
     public List<Member> findByChatroomId(Long chatRoomId) {
